refactor(cli): help, setup, sensitive and the fallback move to src/cli/; main() only wires
Move the remaining inline command blocks out of bin/lcm.ts main() into
src/cli/, verbatim, preserving registration order:

- help command -> src/cli/root.ts, registerHelpCommand(program)
- mcp, install, uninstall -> src/cli/setup.ts, registerSetupCommands(program)
- sensitive -> src/cli/sensitive.ts, registerSensitiveCommand(program)
- unknown command fallback -> src/cli/root.ts, registerUnknownCommandFallback(program)

main() now contains only program setup, the register calls in their
original order, the root help/no-args short-circuit, and parseAsync.
